seperated headers and impl files

+++ src/executor/expressions/number.cpp
@@ -1,8 +1,8 @@
-#include "../executor.hpp"
+#include <executor/executor.hpp>
 
 using NodeType = AST::NodeType; 
 using string = std::string;
 
 std::unique_ptr<Values::Runtime> Interpreter::evaluateNumberLiteral(const AST::NumberLiteral& node){
-	return std::make_unique<Values::Number>(Values::Number(node.value));
+	return std::make_unique<Values::Number>(Values::Number(static_cast<int>(node.value)));
 }
